% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{demo_read_file}
\alias{demo_read_file}
\title{Demo reading a file}
\usage{
demo_read_file(filename)
}
\arguments{
\item{filename}{A filename}
}
\value{
A string
}
\description{
Demo reading a file
}
\examples{
txt <- demo_read_file(system.file("rawdata", "qc.txt", package="rcppstringtest"))
cat(txt)
stopifnot(txt == "Québec\\n" || .Platform$OS.type == "windows")
}
\seealso{
https://fishandwhistle.net/post/2021/using-rs-cross-platform-iconv-wrapper-from-cpp11/
}
