\name{RcppIconvExample-package}
\alias{RcppIconvExample-package}
\alias{RcppIconvExample}
\docType{package}
\title{\packageTitle{RcppIconvExample}}
\description{\packageDescription{RcppIconvExample}}
\section{Package Content}{\packageIndices{RcppIconvExample}}
\author{\packageAuthor{RcppIconvExample}}
\section{Maintainer}{\packageMaintainer{RcppIconvExample}}
\keyword{package}
