\name{rcppstringtest-package}
\alias{rcppstringtest-package}
\alias{rcppstringtest}
\docType{package}
\title{\packageTitle{rcppstringtest}}
\description{\packageDescription{rcppstringtest}}
\section{Package Content}{\packageIndices{rcppstringtest}}
\author{\packageAuthor{rcppstringtest}}
\section{Maintainer}{\packageMaintainer{rcppstringtest}}
\keyword{package}
